// src/types/case-study.ts
// ----------------------------------------------------------------------

export type ICaseStudyProps = {
  id: string;
  title: string;
  heroUrl: string;
  website: string;
  content: string;
  category: string;
  coverUrl: string;
  description: string;
  createdAt: Date;
  galleryImgs: string[];
};